Moving opentracing example
