+++ docs/opentracing/find-traces.png
