+++ docs/opentracing/service-selection.png
